Hiring/requirements.txt
Django==1.7
Flask==0.10.1
Flask-Bootstrap==3.3.5.3
Jinja2==2.7.3
Markdown==2.6.5
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==3.0.0
Py4J==0.6
PyYAML==3.11
SQLAlchemy==1.0.11
Werkzeug==0.10.4
argparse==1.2.1
django-filter==0.12.0
djangorestframework==3.3.2
djoser==0.4.3
dominate==2.1.12
fuzzywuzzy==0.5.0
itsdangerous==0.24
nltk==3.1
numpy==1.10.2
pandas==0.17.1
pbr==1.8.1
progressbar==2.3
pyparsing==2.0.6
python-Levenshtein==0.12.0
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
stevedore==1.10.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
wsgiref==0.1.2
